python/sources/HiveMQ/requirements.txt
quixstreaming==0.4.2a202203161202
paho-mqtt